pyproject.toml
[tool.poetry]
name = "olivka-food-bot"
version = "0.1.0"
description = "Телеграмм бот для olivkafood.ru"
authors = ["Vadim V. <me@vadim7210.ru>", "Timur T."]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9"
python-dotenv = ">=1.0.0"
requests = ">=2.28.2"
aiogram = "=2.25.1"
beautifulsoup4 = ">=4.12.2"
lxml = ">=4.9"
pillow = ">=9.5.0"
loguru = ">=0.7.0"
apscheduler = ">=3.10.1"
emoji = "^2.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
